feat: handle file clashes from different sources

+++ src/main/java/io/github/coenraadhuman/directory/bot/file/manager/SymlinkCreation.java
@@ -1,5 +1,8 @@
 package io.github.coenraadhuman.directory.bot.file.manager;
 
+import io.github.coenraadhuman.directory.bot.configuration.Properties;
+import io.github.coenraadhuman.directory.bot.configuration.Property;
+import io.github.coenraadhuman.directory.bot.utility.Checksum;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -7,16 +10,22 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.time.Instant;
+import java.util.Optional;
+
+import static io.github.coenraadhuman.directory.bot.configuration.Property.DIRECTORY_BOT_SYMLINK_CREATION_AVOID_VALID_OVERWRITE;
 
 public class SymlinkCreation {
 
     private static final Logger log = LoggerFactory.getLogger(SymlinkCreation.class);
-    
+
+    private final Properties properties;
     private final Path sourceDirectory;
     private final Path targetDirectory;
     private final Path sourceFile;
 
-    public SymlinkCreation(Path sourceDirectory, Path targetDirectory, Path sourceFile) {
+    public SymlinkCreation(Properties properties, Path sourceDirectory, Path targetDirectory, Path sourceFile) {
+        this.properties = properties;
         this.sourceDirectory = sourceDirectory;
         this.targetDirectory = targetDirectory;
         this.sourceFile = sourceFile;
@@ -26,6 +35,7 @@ public class SymlinkCreation {
         final var symlinkRootDirectory = Paths.get(sourceFile.getParent().toString().replace(sourceDirectory.toString(), targetDirectory.toString()));
         // Todo: this second argument might undergo a rename if user stipulates it:
         final var symlinkAbsolutePath = Paths.get(symlinkRootDirectory.toString(), sourceFile.getFileName().toString());
+        Optional<Path> clashSymlinkAbsolutePath = Optional.empty();
         
         log.info("Processing source file: {} with determined symlink: {}", sourceFile, symlinkAbsolutePath);
         
@@ -42,22 +52,43 @@ public class SymlinkCreation {
                     log.warn("Invalid symlink: {} found, removing it", symlinkAbsolutePath);
                     Files.delete(symlinkAbsolutePath);
                 } else {
+
                     var existingSymlinkTarget = Files.readSymbolicLink(symlinkAbsolutePath);
                     var existingSymlinkTargetPath = Paths.get(existingSymlinkTarget.getParent().toString(), existingSymlinkTarget.getFileName().toString());
 
                     if (sourceFile.equals(existingSymlinkTargetPath)) {
                         log.info("Skipping creation of symlink: {}, already exists for source file: {}", symlinkAbsolutePath, sourceFile);
                         return;
-                    }
+                    } else {
+                        if (properties.getFlagProperty(DIRECTORY_BOT_SYMLINK_CREATION_AVOID_VALID_OVERWRITE)) {
+                            var newFilePath = Checksum.getMD5(Instant.now().toString())
+                                    .map(checksum -> {
+                                        var filename = symlinkAbsolutePath.getFileName().toString();
+                                        int dotIndex = filename.lastIndexOf('.');
 
-                    // Todo: configuration for how to handle multiple sources with same file name and directory?
-                    log.warn("Symlink: {} with incorrect target, removing it", symlinkAbsolutePath);
-                    Files.delete(symlinkAbsolutePath);
+                                        if (dotIndex == -1) {
+                                            log.warn("File: {} does not have an extension", sourceFile);
+                                            return filename + " - " + checksum;
+                                        }
+
+                                        var extension = filename.substring(dotIndex);
+                                        var baseName = filename.substring(0, dotIndex);
+                                        return baseName + " - " + checksum + extension;
+                                    })
+                                    .map(newFileName -> Paths.get(symlinkRootDirectory.toString(), newFileName))
+                                    .orElseThrow(() -> new RuntimeException("Could not create new name for symlink: {} to avoid overwrite".formatted(symlinkAbsolutePath)));
+
+                            clashSymlinkAbsolutePath = Optional.of(newFilePath);
+                        } else {
+                            log.warn("Symlink: {} with incorrect target, removing it", symlinkAbsolutePath);
+                            Files.delete(symlinkAbsolutePath);
+                        }
+                    }
                 }
             }
 
-            Files.createSymbolicLink(symlinkAbsolutePath, sourceFile);
-            log.info("Created symlink: {} for source file: {}", symlinkAbsolutePath, sourceFile);
+            Files.createSymbolicLink(clashSymlinkAbsolutePath.orElse(symlinkAbsolutePath), sourceFile);
+            log.info("Created symlink: {} for source file: {}", clashSymlinkAbsolutePath.orElse(symlinkAbsolutePath), sourceFile);
         } catch (IOException e) {
             log.error("Could not create symlink: {} for source file: {}", symlinkAbsolutePath, sourceFile);
         }
